pyproject.toml
[tool.poetry]
name = "firehole"
version = "0.1.0"
description = ""
authors = [
    "Jiří Rája <raja@ics.muni.cz>",
    "Daniel Volf <248688@mail.muni.cz>",
]
maintainers = [
    "Jiří Rája <raja@ics.muni.cz>",
    "Daniel Volf <248688@mail.muni.cz>",
]
readme = "README.md"
packages = [{include = "firehole", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[tool.poetry.scripts]
firehole = 'firehole.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
